requirements.txt
asgiref==3.5.2
Brotli==1.0.9
dj-database-url==2.0.0
Django==4.0.4
django-cors-headers==4.1.0
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
importlib-metadata==5.0.0
inflection==0.5.1
Markdown==3.4.1
mysqlclient==2.1.1
packaging==23.1
psycopg2==2.9.3
psycopg2-binary==2.9.6
pytz==2022.4
PyYAML==6.0
sqlparse==0.4.3
typing_extensions==4.6.3
tzdata==2022.4
uritemplate==4.1.1
whitenoise==6.5.0
zipp==3.8.1
